build: Update renovate config.
Update the renovate config in edx-platform to match the standard config
in most of our other frontend repos, with the exception of 1 settings:

Given the size of edx-platform I want to limit the number of concurrent
renovate PRs to a very small number so that we don't overwhelm our CI
resources.  If we find that it would be useful to get more PRs at a
time, we can change this setting in the future.